A little more cleanup and updating on models
Update on the walls models, using the lighter color one.
Update on the human player model.
Updated prefabs, locker, flashlight, crowbar.
Temporary fix to the floating problem while walking.
Changed the "Body" script so the severed limbs have the same material as the ones on the origin.
Added the limb cap texture, and updated the color on the player material.
Added a simple collision layer, Items, that just avoid items colliding with each other.