Asset Update
Rotation fixed on the chairs.
Updated all the floor meshes to the one with 2 tris.
Updated the tables rotation and little change on the color of one table.
Updated the boombox rotation and some fixes on the mesh.
Fixes on the rotation and mesh of the extinguisher.
Update on the opacity of Palette05 Opacity.
Texture changes from filter to point.
Created prefab for the plastic char blue.
Updated locker shadows.
Visual changes on the materials.
I also changed some quality changes on the different graphic tiers