Movement and UI fixes
Fixed all the floor prefabs to have convex enabled on the mesh collider.
All the UI changes were Pixel Per Unit on 3 to add more quality to the panels.
The HumanoidMovementController now has a little better IK and more control over the rotations, aswell as fixing the bug with walking, stopping, and pressing Q or E on the same frame causing the character to move towards that position, also fixed the head rotation caused by that same thing.
Changed the DelayedInteraction's loading bar spawn to up because it spawns on the ground.