Simple radial menu for interactions
This adapts Onebi's radial menu system to work with interactions, currently the interactions have the missing item icon, but they work.
Many errors that already occured before might happen again, as null references when you click on something that is not the radial menu.
Also UI scaling is not done yet.