Fully functional radial menu
adds the funcionality of a radial menu for interactions, clicking on a interactable object creates the radial menu with it's interactions and the objects in hand interactions (which should be reworked with Alain's updates).
adds some icons for those interactions.
fixes some animations of the radial menu.
adds two fonts.
alterations on prefabs for the interaction icons while we can't put them into a base object.