Radial menu size changer and bug fixes
you no longer open a blank menu when clicking the UI slot.
added a size range for the menu.
added a "objectName" for the menu in case we want to debug it.
changed the item pickups to the item's icon.
fixed font related issues as sizing and compression.
added the Z again for activating items.