Create PepperSpray class to handle interaction
Fixes #600, adds a class that plays the spray effect when the item is interacted with. Ideally, interacting with the item should cycle between power like in the normal game (in fact, as all spray items have this power feature, they should probably share an abstract class or interface), while clicking somewhere would fire in that direction, but no combat system so whatever.