Update database
Added registered users to the database so the server knows which ckeys have logged in. Also added a username panel to test that, shouldn't be there once we have a good method of getting users.
Added a simple job selection panel because I only realized I needed the ckey after starting it.