Changed treatment of Examinables where netId = 0. Previously, they had been treated like non-networked objects, with corresponding text displayed. They now display an error message in lieu of the standard Examine message. This is to ensure that network issues are not hidden by functionality that looks like it works.
Reworked networking interactions - previously, networked items were performing the checks twice on the client and not at all on the server. Code not yet working properly - commands are failing in the server.

Added debugging comments into Examinator, to attempt to find out why client-initiated Examine requests are failing in the server.