Created IExamineData interface and ExamineHelper class.
The purpose of the IExamineData interface is to package different data fields from the examined object, so that they can be passed to a compatible user interface.

The purpose of the ExamineHelper class is to manage the interaction between data and user interfaces. It contains the enum ExamineType, which is used to match these.