Attempted to allow clients to load map and embark on round. **Known issues exist**
As part of Mirror update, MessagePacking (formerly MessagePacker) now disconnects connections which have message handler exceptions. This was causing the client to immediately return to the main menu after clicking "Embark". I have disabled this auto-disconnect for the moment, but I still do not understand the root cause. This requires further investigation.

SceneLoaderManager now asks the client to load the map when the server does, rather than wait for the server to finish. This prevents errors associated with spawning scene objects before the PrepareToSpawnSceneObjects method is called.

RoundManager amended to prevent clients from attempting to perform actions that only the server should be able to.

**Known issues**: On clients, the orientation of some airlocks is incorrect. Additionally, some RPC's are failing because the network identity has already been destroyed. These issues may be resolved by the tilemap rework.