Simplified the StartRound() and StartWarmup() functions.
Above functions were amended to separate the client-side actions from the server-side actions. This allows the RPC's to simply call the primary method, so all future changes are in one place.

Additionally, this fix prevents the issue where clients load the Tick and TickWarmup coroutines and illegally attempt to update the server clock.