CompositeItemSelector.cs no longer records the mesh texture.
Upon review of the code, it was clear that the recorded texture was never actually used anywhere. Additionally, by allocating the texture to the singleColourMaterial, it was changing that file every time the Examine system was used. This caused the file to want to update within version control after every commit.