Updated LoginNetworkManager to more clearly define Client and Server methods.
SpawnPlayerAfterRoundStart() is a server-only method. Previously, it called SpawnPlayerWithoutLoginServer(), which should be a client-only method because it sends a CharacterSelectMessage. CharacterSelectMessages should only be sent from client to server. Because it was being sent from server to client instead, it was causing errors as there was not a registered handler of these messages on the client.