Removes containerSync on object when not needed
Some gameObject need a containerSync to work, such as a cardboardbox.
But when the cardboardbox is a child of another gameObject with a containerSync, there should be only one containerSync for the whole hierarchy, parent and childs. This insures us that when adding a new gameObject to the scene, only one containerSync is on the object.