Added NetworkStartPositions to MiniStation.
A somewhat disturbing bug had emerged where the Human prefab looked monstrous on the server (though looked normal on the client). It appeared that this was caused by the players instantiating in exactly the same position, as it did not occur when the players had moved away prior to the next one embarking. The four NetworkStartPositions added ensures that players will not routinely spawn as monsters.

It is possible that players *might* still look monstrous if they move onto another spawn point as a new player spawns there, but I think that is unlikely enough not to worry about for the moment.