Refactored Gamemode system to separate domain layer from service layer.
GamemodeSystem is the service layer, and handles all networking, cross-system communication etc. It collects and passes in the data required by the domain layer. Potentially rename in future to "GamemodeController".

Gamemode, GamemodeObjective and remainder form the domain layer, and do not (and should not) have knowledge of the service layer, external systems, networking etc.