Greatly simplifies  setting container
The old way was very prone to break, I guess it was done for performance reasons, but you had to send a bool to say if it was already in the container or already removed. Here, I simply check that in SetContainer. Slower but sturdy. We'll optimise on that eventually. Also remove the dangerous "setting container unchecked" method. Surely efficient sometimes, but it's public and very prone to break too. That's end game optimisation.