add human entity and some fix
got some issues ... when head is cut, a head item spawn and there's a mind swap on the new entity which works okay. However this "new" entity (the head), doesn't register in the entity system and is not found by it. I wanted to simply add it to the _spawnedPlayers synclist but it triggers a bunch of other stuff in the game mode system and such and I'm unsure if it's any wanted behavior ... fixed by setting properly the synclist in entitySystem