Container simplification (#1243)
* remove container script

* simplify some more

* simplify some more

* remove size of items in container

* readd check if area is free

* fix moving items on grid

Items used to be dropped at the center of their sprite, it's more intuitive to do it at mouse position. Also fix a bug with sprite when dropped on the edge of the container display.

* only remove item if necessary

no need to remove item from the container if it's moved inside said container

* remove useless debug

* fix item sprite glitch

fix the annoying glitch where the item appears for a frame at the old place it was. This simply hide the item sprite until it's transferred.

* add dump container command for tests

* some comments

* add init and fix item being able to be contained in overcharged containers

* remove size from item init

* fix container tests and remove useless tests

Some tests are only linked to item size, they are no longer useful

* remove useless member as it was only used for 2d items sprite

* fix custom editor of containers

* simplify call to event

Since event is only called by synclist callback, only one item at a time need to be dealt with, including when calling clear on the synclist

* make canhold and can store private

* update conflicts