Fix comparison between traits not working correctly
This is a weird issue that only happens on the client but not on the host. The type of the same two traits returns a different value, the first trait that is getting compared returns the general/main object type (Trait) while the second returns the specific type. For example, Security trait returns "Trait" object type for item 1 being compared, while the same trait returns "IDPermission" object type for item 2